backend/requirements.txt
flask
flask-cors
ultralytics
google-generativeai
python-dotenv
opencv-python-headless
Pillow
numpy
requests